requirements.txt
streamlit==1.55.0
psycopg2-binary==2.9.9
pandas==3.0.1
google-cloud-aiplatform>=1.80.0
python-dotenv==1.0.1
